Module people_rococo_integration_tests::xcm_helpers

Helpers for calculating XCM delivery fees.

Functions

pay_over_xcm_delivery_fees

Returns the delivery fees amount for the execution of PayOverXcm

query_response_delivery_fees

Returns the delivery fees amount for a query response as a result of the execution of a ExpectError instruction with no error.

transfer_assets_delivery_fees

Returns the delivery fees amount for pallet xcm’s teleport_assets and reserve_transfer_assets extrinsics. Because it returns only a u128, it assumes delivery fees are only paid in one asset and that asset is known.